Fix labels filter popover overflow and surface create affordance
